+++ e2e/e2etests/test_eth_withdraw_restricted_address.go
@@ -3,6 +3,7 @@ package e2etests
 import (
 	"math/big"
 
+	"cosmossdk.io/math"
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	ethcommon "github.com/ethereum/go-ethereum/common"
 	"github.com/stretchr/testify/require"
@@ -10,7 +11,9 @@ import (
 
 	"github.com/zeta-chain/node/e2e/runner"
 	"github.com/zeta-chain/node/e2e/utils"
+	crosschainkeeper "github.com/zeta-chain/node/x/crosschain/keeper"
 	crosschaintypes "github.com/zeta-chain/node/x/crosschain/types"
+	observertypes "github.com/zeta-chain/node/x/observer/types"
 )
 
 // TestEtherWithdrawRestricted tests the withdrawal to a restricted receiver address
@@ -71,5 +74,39 @@ func TestEtherWithdrawRestricted(r *runner.E2ERunner, args []string) {
 	// revert address should receive the amount
 	revertBalanceAfter, err := r.ETHZRC20.BalanceOf(&bind.CallOpts{}, revertAddress)
 	require.NoError(r, err)
-	require.EqualValues(r, new(big.Int).Add(revertBalanceBefore, amount), revertBalanceAfter)
+
+	userBalanceAfterUint := math.NewUintFromBigInt(revertBalanceAfter)
+	userBalanceBeforeUint := math.NewUintFromBigInt(revertBalanceBefore)
+	totalRevertAmount := getTotalRevertedAmount(r, cctx)
+
+	require.EqualValues(r, userBalanceAfterUint.Sub(totalRevertAmount), userBalanceBeforeUint)
+}
+
+func getTotalRevertedAmount(r *runner.E2ERunner, cctx *crosschaintypes.CrossChainTx) math.Uint {
+	outboundParams := cctx.OutboundParams[0]
+	outboundTxGasUsed := math.NewUint(outboundParams.GasUsed)
+	outboundTxFinalGasPrice := math.NewUintFromBigInt(outboundParams.EffectiveGasPrice.BigInt())
+	outboundTxFeePaid := outboundTxGasUsed.Mul(outboundTxFinalGasPrice)
+	userGasFeePaid := outboundParams.UserGasFeePaid
+	totalRemainingFees := userGasFeePaid.Sub(outboundTxFeePaid)
+
+	remainingFees := crosschainkeeper.PercentOf(totalRemainingFees, crosschaintypes.UsableRemainingFeesPercentage)
+	if remainingFees.LTE(math.ZeroUint()) {
+		return math.ZeroUint()
+	}
+
+	evmChainID, err := r.EVMClient.ChainID(r.Ctx)
+	require.NoError(r, err)
+
+	chainParams, err := r.ObserverClient.GetChainParamsForChain(
+		r.Ctx,
+		&observertypes.QueryGetChainParamsForChainRequest{
+			ChainId: evmChainID.Int64(),
+		},
+	)
+	require.NoError(r, err)
+
+	stabilityPoolAmount := crosschainkeeper.PercentOf(remainingFees, chainParams.ChainParams.StabilityPoolPercentage)
+	refundAmount := remainingFees.Sub(stabilityPoolAmount)
+	return cctx.InboundParams.Amount.Add(refundAmount)
 }