+++ x/crosschain/keeper/msg_server_vote_outbound.go
@@ -3,16 +3,18 @@ package keeper
 import (
 	"context"
 	"fmt"
-	"math/big"
 
 	cosmoserrors "cosmossdk.io/errors"
 	"cosmossdk.io/math"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
+	ethcommon "github.com/ethereum/go-ethereum/common"
+	"github.com/pkg/errors"
 
 	cctxerror "github.com/zeta-chain/node/pkg/errors"
 	"github.com/zeta-chain/node/x/crosschain/types"
 	observerkeeper "github.com/zeta-chain/node/x/observer/keeper"
+	observertypes "github.com/zeta-chain/node/x/observer/types"
 )
 
 const voteOutboundID = "Vote Outbound"
@@ -112,14 +114,13 @@ func (k msgServer) VoteOutbound(
 	// Set the finalized ballot to the current outbound params.
 	cctx.SetOutboundBallotIndex(ballotIndex)
 
-	// If ballot is successful, the value received should be the out tx amount.
-	err = cctx.AddOutbound(ctx, *msg, ballot.BallotStatus)
+	err = cctx.UpdateCurrentOutbound(ctx, *msg, ballot.BallotStatus)
 	if err != nil {
 		return nil, cosmoserrors.Wrap(err, voteOutboundID)
 	}
 
 	// Fund the gas stability pool with the remaining funds.
-	k.FundStabilityPool(ctx, &cctx)
+	k.RefundUnusedGasFee(ctx, &cctx)
 
 	// Validate and process the observed outbound
 	err = k.ValidateOutboundObservers(ctx, &cctx, ballot.BallotStatus, msg.ValueReceived.String())
@@ -136,19 +137,108 @@ func (k msgServer) VoteOutbound(
 	return &types.MsgVoteOutboundResponse{}, nil
 }
 
-// FundStabilityPool funds the stability pool with the remaining fees of an outbound tx
-// The funds are sent to the gas stability pool associated with the receiver chain
-// This wraps the FundGasStabilityPoolFromRemainingFees function and logs an error if it fails.We do not return an error here.
+// RefundUnusedGasFee funds the stability pool with the remaining fees of an outbound tx
+// The funds are sent to the gas stability pool associated with the receiver chain and the user is refunded if possible.
+// This wraps the RefundUnusedGas function and logs an error if it fails.We do not return an error here.
 // Event if the funding fails, the outbound tx is still processed.
-func (k Keeper) FundStabilityPool(ctx sdk.Context, cctx *types.CrossChainTx) {
-	// Fund the gas stability pool with the remaining funds
-	if err := k.FundGasStabilityPoolFromRemainingFees(ctx, *cctx.GetCurrentOutboundParam(), cctx.GetCurrentOutboundParam().ReceiverChainId); err != nil {
-		ctx.Logger().
-			Error("%s: CCTX: %s Can't fund the gas stability pool with remaining fees %s", voteOutboundID, cctx.Index, err.Error())
+func (k Keeper) RefundUnusedGasFee(ctx sdk.Context, cctx *types.CrossChainTx) {
+	outboundParams := cctx.GetCurrentOutboundParam()
+	// We skip funding the gas stability pool if the userGasFeePaid is nil or zero.
+	// This is a legacy outbound where the user fee was not recorded as part of the cctx struct. This is only to handle cctxs which might be in pending outbound state when the upgrade happens and get finalized after the upgrade
+	if outboundParams.UserGasFeePaid.IsNil() {
+		err := k.FundGasStabilityPoolFromRemainingFees(
+			ctx,
+			*outboundParams,
+			outboundParams.ReceiverChainId,
+		)
+		if err != nil {
+			ctx.Logger().Error("Failed to fund gas stability pool (Legacy) with remaining fees",
+				"voteOutboundID", voteOutboundID,
+				"cctxIndex", cctx.Index,
+				"error", err,
+			)
+		}
+		return
+	}
+
+	if err := k.refundUnusedGas(ctx, cctx); err != nil {
+		ctx.Logger().Error("Failed to fund gas stability pool with remaining fees",
+			"voteOutboundID", voteOutboundID,
+			"cctxIndex", cctx.Index,
+			"error", err,
+		)
 	}
 }
 
+// RefundUnusedGas uses the remaining fees of an outbound tx to fund the gas stability pool and refund the user if possible
+func (k Keeper) refundUnusedGas(
+	ctx sdk.Context,
+	cctx *types.CrossChainTx,
+) error {
+	outboundParams := cctx.GetCurrentOutboundParam()
+	outboundTxFeePaid := math.NewUint(outboundParams.GasUsed).
+		Mul(math.NewUintFromBigInt(outboundParams.EffectiveGasPrice.BigInt()))
+	userGasFeePaid := outboundParams.UserGasFeePaid
+
+	// Handle cases for not funding the stability pool or refunding the user
+	// 1. If outboundTxFeePaid is nil or zero, we do not fund the stability pool or refund the user.This is for Non EVM chains which do not populate outboundTxFeePaid
+	// 2. If outboundTxFeePaid is greater than or equal to userGasFeePaid, we do not fund the stability pool or refund the user. Since the outbound tx used all the gas paid by the user.The additional gas used is covered by the stability pool.
+	// https://github.com/zeta-chain/node/issues/4219
+	// Enable for non EVM chains once zeta-client supports populating gas used and effective gas price for non EVM chains.
+	if outboundTxFeePaid.IsNil() || outboundTxFeePaid.IsZero() || outboundTxFeePaid.GTE(userGasFeePaid) {
+		return nil
+	}
+
+	// We use a maximum for 95 percent of the remaining fees to fund the stability pool and refund the user.
+	usableRemainingFees := PercentOf(
+		outboundParams.UserGasFeePaid.Sub(outboundTxFeePaid),
+		types.UsableRemainingFeesPercentage,
+	)
+	if !usableRemainingFees.GT(math.ZeroUint()) {
+		return nil
+	}
+
+	// Send all tokens to stability pool by default
+	stabilityPoolPercentage := types.DefaultStabilityPoolFundPercentage
+	refundToUser := false
+
+	isWithdrawTx, err := cctx.IsWithdrawTx()
+	if err != nil {
+		return errors.Wrap(err, "failed to determine if the tx is a withdrawal")
+	}
+
+	// Refund to the sender irrespective of whether its EOA or contract address if it's a withdrawal originating from zEVM.
+	if isWithdrawTx && ethcommon.IsHexAddress(cctx.InboundParams.Sender) {
+		chainParams, found := k.GetObserverKeeper().GetChainParamsByChainID(ctx, outboundParams.ReceiverChainId)
+		if !found {
+			return errors.Wrap(
+				observertypes.ErrChainParamsNotFound,
+				fmt.Sprintf("chainID: %d", outboundParams.ReceiverChainId),
+			)
+		}
+		refundToUser = true
+		stabilityPoolPercentage = chainParams.StabilityPoolPercentage
+	}
+	stabilityPoolAmount := PercentOf(usableRemainingFees, stabilityPoolPercentage)
+	if stabilityPoolAmount.GT(math.ZeroUint()) {
+		if err := k.fungibleKeeper.FundGasStabilityPool(ctx, outboundParams.ReceiverChainId, stabilityPoolAmount.BigInt()); err != nil {
+			return err
+		}
+	}
+
+	refundAmount := usableRemainingFees.Sub(stabilityPoolAmount)
+	if refundAmount.GT(math.ZeroUint()) && refundToUser {
+		if err := k.fungibleKeeper.DepositChainGasToken(ctx, outboundParams.ReceiverChainId, refundAmount.BigInt(), ethcommon.HexToAddress(cctx.InboundParams.Sender)); err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
 // FundGasStabilityPoolFromRemainingFees funds the gas stability pool with the remaining fees of an outbound tx
+// TODO: Remove this function
+// This function handles the legacy flow for funcding the gas stability pool with the remaining fees.
 func (k Keeper) FundGasStabilityPoolFromRemainingFees(
 	ctx sdk.Context,
 	OutboundParams types.OutboundParams,
@@ -166,13 +256,13 @@ func (k Keeper) FundGasStabilityPoolFromRemainingFees(
 	if gasLimit > 0 && gasUsed > 0 && !gasPrice.IsZero() {
 		if gasLimit > gasUsed {
 			remainingGas := gasLimit - gasUsed
-			remainingFees := math.NewUint(remainingGas).Mul(gasPrice).BigInt()
+			remainingFees := math.NewUint(remainingGas).Mul(gasPrice)
 
 			// We fund the stability pool with a portion of the remaining fees.
-			remainingFees = percentOf(remainingFees, RemainingFeesToStabilityPoolPercent)
+			remainingFees = PercentOf(remainingFees, RemainingFeesToStabilityPoolPercent)
 
 			// Fund the gas stability pool.
-			if err := k.fungibleKeeper.FundGasStabilityPool(ctx, chainID, remainingFees); err != nil {
+			if err := k.fungibleKeeper.FundGasStabilityPool(ctx, chainID, remainingFees.BigInt()); err != nil {
 				return err
 			}
 		} else {
@@ -183,11 +273,11 @@ func (k Keeper) FundGasStabilityPoolFromRemainingFees(
 	return nil
 }
 
-// percentOf returns the percentage of a number
-func percentOf(n *big.Int, percent int64) *big.Int {
-	n = n.Mul(n, big.NewInt(percent))
-	n = n.Div(n, big.NewInt(100))
-	return n
+// PercentOf returns the percentage of a number
+func PercentOf(n math.Uint, percent uint64) math.Uint {
+	percentUint := math.NewUint(percent)
+	result := n.Mul(percentUint)
+	return result.Quo(math.NewUint(100))
 }
 
 // SaveOutbound saves the outbound transaction.It does the following things in one function: