+++ x/fungible/keeper/evm_test.go
@@ -20,6 +20,7 @@ import (
 	"github.com/zeta-chain/protocol-contracts/pkg/wzeta.sol"
 	"github.com/zeta-chain/protocol-contracts/pkg/zrc20.sol"
 
+	ethcommon "github.com/ethereum/go-ethereum/common"
 	"github.com/zeta-chain/node/e2e/contracts/dapp"
 	"github.com/zeta-chain/node/e2e/contracts/dappreverter"
 	"github.com/zeta-chain/node/e2e/contracts/example"
@@ -1650,5 +1651,45 @@ func TestKeeper_CallOnRevertZevmConnector(t *testing.T) {
 			big.NewInt(1),
 			big.NewInt(45), []byte("message"), [32]byte{})
 		require.ErrorContains(t, err, "execution reverted")
+	})
+}
+
+func TestKeeper_RefundRemainingGasFees(t *testing.T) {
+	t.Run("should error if system contracts not deployed", func(t *testing.T) {
+		k, ctx, _, _ := keepertest.FungibleKeeper(t)
+		_ = k.GetAuthKeeper().GetModuleAccount(ctx, types.ModuleName)
+
+		chainID := getValidChainID(t)
+		receiver := ethcommon.HexToAddress("0x742d35Cc6634C0532925a3b844Bc454e4438f44e")
+
+		err := k.DepositChainGasToken(ctx, chainID, big.NewInt(100), receiver)
+		require.Error(t, err)
+	})
+
+	t.Run("can refund remaining gas fees to receiver", func(t *testing.T) {
+		// Arrange
+		k, ctx, sdkk, _ := keepertest.FungibleKeeper(t)
+		_ = k.GetAuthKeeper().GetModuleAccount(ctx, types.ModuleName)
+
+		chainID := getValidChainID(t)
+		receiver := ethcommon.HexToAddress("0x742d35Cc6634C0532925a3b844Bc454e4438f44e")
+
+		deploySystemContracts(t, ctx, k, sdkk.EvmKeeper)
+		gasZRC20 := setupGasCoin(t, ctx, k, sdkk.EvmKeeper, chainID, "foobar", "foobar")
+
+		receiverBalanceBefore, err := k.BalanceOfZRC4(ctx, gasZRC20, receiver)
+		require.NoError(t, err)
+		require.Equal(t, int64(0), receiverBalanceBefore.Int64())
+
+		// Act
+		err = k.DepositChainGasToken(ctx, chainID, big.NewInt(100), receiver)
+		require.NoError(t, err)
+
+		// Assert
+		gasZRC20, err = k.QuerySystemContractGasCoinZRC20(ctx, big.NewInt(chainID))
+		require.NoError(t, err)
+		receiverBalanceAfter, err := k.BalanceOfZRC4(ctx, gasZRC20, receiver)
+		require.NoError(t, err)
+		require.Equal(t, int64(100), receiverBalanceAfter.Int64())
 	})
 }