+++ x/observer/types/chain_params.go
@@ -141,6 +141,13 @@ func (cp ChainParams) Validate() error {
 		return ErrParamsMinObserverDelegation
 	}
 
+	if cp.StabilityPoolPercentage > 100 {
+		return errors.Wrapf(
+			ErrParamsStabilityPoolPercentage,
+			"stability pool percentage must be in range [0,100], got: %d",
+			cp.StabilityPoolPercentage,
+		)
+	}
 	return nil
 }