+++ changelog.md
@@ -18,6 +18,7 @@ by calling `updateAdditionalActionFee` admin function.
 * [4157](https://github.com/zeta-chain/node/pull/4157) - multiple evm calls in single tx
 * [4211](https://github.com/zeta-chain/node/pull/4211) - provide error information in cctx when Bitcoin deposit fail
 * [4218](https://github.com/zeta-chain/node/pull/4218) - enable NoAssetCall from Bitcoin chain
+* [3834](https://github.com/zeta-chain/node/pull/3734) - refund a portion of remaining unused tokens to user
 
 ### Refactor